+++ app/src/main/java/de/baumann/browser/Browser_left.java
@@ -301,8 +301,9 @@ public class Browser_left extends AppCompatActivity implements ObservableScrollV
         } else if (Intent.ACTION_VIEW.equals(action)) {
             new Handler().postDelayed(new Runnable() {
                 public void run() {
-                    String sharedText = intent.getStringExtra(Intent.EXTRA_TEXT);
-                    mWebView.loadUrl(sharedText);
+                    Uri data = intent.getData();
+                    String link = data.toString();
+                    mWebView.loadUrl(link);
                 }
             }, 300);
         } else if ("closeAPP".equals(action)) {
@@ -890,7 +891,7 @@ public class Browser_left extends AppCompatActivity implements ObservableScrollV
                                 if(db.isExist(mWebView.getUrl())){
                                     Snackbar.make(editText, getString(R.string.toast_newTitle), Snackbar.LENGTH_LONG).show();
                                 }else{
-                                    db.insert(mWebView.getTitle(), mWebView.getUrl(), "", "", helper_main.createDate());
+                                    db.insert(helper_webView.getTitle (mWebView), mWebView.getUrl(), "", "", helper_main.createDate());
                                     Snackbar.make(mWebView, R.string.bookmark_added, Snackbar.LENGTH_LONG).show();
                                 }
                             }